refactor: extract SpeakerSocialLinks and align icons to flex-start
