+++ app/eventyay/webapp/schedule/src/components/SpeakersList.vue
@@ -132,18 +132,7 @@
 								h4 {{ speaker.name || t.speaker_fallback }}
 								markdown-content.featured-speaker-preview-bio(v-if="speaker.biography", :markdown="speaker.biography")
 					.featured-speaker-details
-						.speaker-social-links(v-if="speaker.social_links && speaker.social_links.length")
-							a.speaker-social-link(
-								v-for="link in speaker.social_links",
-								:key="link.key + link.url",
-								:href="link.url",
-								:class="'speaker-social-link--' + link.key",
-								:style="{ color: link.color || undefined }",
-								:aria-label="link.label",
-								:title="link.label",
-								target="_blank",
-								rel="noopener noreferrer")
-								span.speaker-social-svg(v-html="getSocialIconHtml(link)")
+						speaker-social-links(:links="speaker.social_links", alignment="flex-start")
 						template(v-if="speaker.sessions && speaker.sessions.length")
 							hr.featured-speaker-divider(v-if="speaker.social_links && speaker.social_links.length")
 							hr.featured-speaker-divider(v-else)
@@ -167,8 +156,9 @@
 
 <script>
 import moment from 'moment-timezone'
-import { getLocalizedString, compareFeaturedSpeakers, isFeaturedSpeakersSortAvailable, sessionsForSpeaker, getSocialIconHtml } from '../utils'
+import { getLocalizedString, compareFeaturedSpeakers, isFeaturedSpeakersSortAvailable, sessionsForSpeaker } from '../utils'
 import MarkdownContent from './MarkdownContent'
+import SpeakerSocialLinks from './SpeakerSocialLinks.vue'
 
 function normalizeLocaleCode (code) {
 	if (!code || typeof code !== 'string') return null
@@ -191,7 +181,7 @@ function localesMatch (filterValue, sessionValue) {
 
 export default {
 	name: 'SpeakersList',
-	components: { MarkdownContent },
+	components: { MarkdownContent, SpeakerSocialLinks },
 	inject: {
 		scheduleData: { default: null },
 		eventUrl: { default: '' },
@@ -230,7 +220,6 @@ export default {
 	data() {
 		return {
 			getLocalizedString,
-			getSocialIconHtml,
 			searchQuery: '',
 			selectedLanguages: [],
 			selectedTracks: [],
@@ -823,42 +812,6 @@ export default {
 			padding: 12px
 			background: $clr-grey-100
 			border-top: 1px solid $clr-grey-300
-
-		.speaker-social-links
-			display: flex
-			flex-wrap: wrap
-			gap: 8px
-			margin-bottom: 4px
-			justify-content: center
-
-		.speaker-social-link
-			display: inline-flex
-			align-items: center
-			justify-content: center
-			width: 32px
-			height: 32px
-			border-radius: 6px
-			background: rgba(0, 0, 0, 0.06)
-			color: inherit
-			text-decoration: none
-			font-size: 16px
-			transition: background-color 0.15s ease, transform 0.15s ease
-			&:hover, &:focus-visible
-				background: rgba(0, 0, 0, 0.12)
-				transform: translateY(-2px)
-			&:active
-				transform: translateY(0)
-
-		.speaker-social-svg
-			display: flex
-			align-items: center
-			justify-content: center
-			width: 20px
-			height: 20px
-			svg
-				width: 100%
-				height: 100%
-				fill: currentColor
 
 		.featured-speaker-divider
 			margin: 12px 0 8px
